edit_buys_funtoin
商品の情報をもとにクレジットカードで購入できるように変更
クレジットカードの一覧表示、登録、削除ができるよう変更
