+++ Gemfile
@@ -85,3 +85,4 @@ gem "payjp"
 gem "carrierwave"
 gem "active_hash"
 gem "enum_help"
+gem "rails-i18n"
