+++ app/views/api/cards/create.html.haml
@@ -1,1 +1,21 @@
-カードの登録が完了しました！
+
+.main
+  .header
+    = link_to root_path do
+      = image_tag 'logo/logo.png', size: "185x49", class: "header__logo"
+  .main__container
+    .main__container__form
+      .content-wrap
+        .complete-text
+          %p カードの登録が完了いたしました！
+          %p{style: "width:360px;"} 引き続きショッピングをお楽しみください！
+        %a{ href: accounts_path(current_user.id), class: "btn" } マイページへ戻る
+    .image-box
+  .footer
+    .footer__lists
+      %a{ href: '#' } プライバシーポリシー
+      %a{ href: '#' } 利用規約
+      %a{ href: '#' } 特定商取引に関する表記
+    %a{ href: '/' }
+      = image_tag 'logo/logo.png', size: "185x49", class: "footer__logo"
+    %p ©FURIMA
