+++ app/views/api/cards/index.html.haml
@@ -1,0 +1,14 @@
+
+.api_cards_index{class:"api_cards_new"}
+  %h2 登録カード一覧
+  - @customer.cards.each do |card|
+    .buys_new-contents-block
+      .buys_new-contents-method
+        .buys_new-contents-method-top
+          - if @customer.default_card == card.id 
+            %p デフォルトカード 
+          =image_tag "card_logo/#{card.brand}.svg"
+          %p="**********#{card.last4}"
+          %p="有効期限 #{card.exp_month}/#{card.exp_year}" 
+        %p= link_to "削除", api_card_path(current_user.cards.find_by(card_id:card.id).id) , method: :delete
+
