+++ app/views/buys/create.html.haml
@@ -7,8 +7,9 @@
     .main__container__form
       .content-wrap
         .complete-text
-          %p カード情報を削除いたしました！
-        %a{ href: accounts_path(current_user.id), class: "btn" } マイページへ戻る
+          %p 購入手続きが完了いたしました！
+          %p{style: "width:360px;"} 出品者からの連絡をお待ちください。
+        %a{ href: root_path, class: "btn" } トップページへ戻る
     .image-box
   .footer
     .footer__lists
