+++ app/views/buys/new.html.haml
@@ -1,0 +1,47 @@
+
+.buys_new
+  = image_tag "logo/logo.png"
+  .buys_new-contents
+    %h2 購入内容の確認
+    .buys_new-contents-block
+      .buys_new-contents-product
+        =image_tag @item.images.first.image.url
+        .items-index-contents-product-sentence
+          %p=@item.name
+          %p=@item.status
+          %p=@item.post_date
+    .buys_new-contents-block
+      .buys_new-contents-price
+        .buys_new-contents-price-top
+          %p 支払い金額
+        %p="#{@item.price} 円"
+    .buys_new-contents-block
+      .buys_new-contents-method
+        .buys_new-contents-method-top
+          %p 支払い方法
+          %p クレジットカード
+          %p="**********#{@card_data.last4}"
+          %p="有効期限 #{@card_data.exp_month}/#{@card_data.exp_year}"
+          =image_tag "card_logo/#{@card_data.brand}.svg"
+    .buys_new-contents-block
+      .buys_new-contents-post
+        .buys_new-contents-post-top
+          %p 配送先
+          %p="〒#{@address.zipcode.insert(3, " - ")}"
+          %p="#{@address.prefecture.name}#{@address.city}#{@address.build_name}"
+          %p="#{current_user.first_name} #{current_user.last_name}"
+    - if current_user.cards.blank?
+      %h2 カードを登録してから、ご購入ください。
+    - else
+      .buys_new-contents-block
+        = button_to "購入する" , item_buys_path(@item.id), method: :post
+  .buys_new-fotter
+    .buys_new-fotter-lists
+      %p プライバシーポリシー
+      %p 利用企画
+      %p 特定商取引に関する表記
+    = image_tag "logo/logo.png"
+    %p &copy; FURIMA
+    
+
+
