--- app/views/items/buys/create.html.haml
@@ -1,1 +1,0 @@
-購入が完了しました！
