--- app/views/items/buys/new.html.haml
@@ -1,50 +1,0 @@
-
-.items_buys_new
-  = image_tag "logo/logo.png"
-  .items_buys_new-contents
-    %h2 購入内容の確認
-    .items_buys_new-contents-block
-      .items_buys_new-contents-product 
-        .items_buys_new-contents-product-img
-        .items-index-contents-product-sentence
-          %p 商品名
-          %p 商品の状態
-          %p 発送の日時
-    .items_buys_new-contents-block
-      .items_buys_new-contents-price
-        .items_buys_new-contents-price-top
-          %p 支払い金額
-          %p ポイント使用（所持ポイント：P0）
-        %p 5000円
-    .items_buys_new-contents-block
-      .items_buys_new-contents-method
-        .items_buys_new-contents-method-top
-          %p 支払い方法
-          %p クレジットカード
-          %p **********999
-          %p 有効期限 09/20
-          .items_buys_new-contents-method-img
-        %p= link_to "変更する", "#"
-    .items_buys_new-contents-block
-      .items_buys_new-contents-post
-        .items_buys_new-contents-post-top
-          %p 配送先
-          %p 〒910-0315
-          %p 詳しい住所
-          %p 辻村　真琴
-        %p= link_to "変更する", "#"
-    - if current_user.cards.blank?
-      %h2 カードを登録してから、ご購入ください。
-    - else
-      .items_buys_new-contents-block
-        = button_to "購入する" , items_buys_path, method: :post
-  .items_buys_new-fotter
-    .items_buys_new-fotter-lists
-      %p プライバシーポリシー
-      %p 利用企画
-      %p 特定商取引に関する表記
-    = image_tag "logo/logo.png"
-    %p &copy; FURIMA
-    
-
-
